Inintial phase

+++ src/main/java/com/example/demos/Repository/foodrepository.java
@@ -3,5 +3,5 @@ package com.example.demos.Repository;
 import com.example.demos.entity.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface foodrepository extends JpaRepository<Food, Integer> {
+public interface foodrepository extends JpaRepository<Food, Long> {
 }
